Integration test framework.
We need a consistent service to properly test this library.

This commit adds a small sinatra app that contains endpoints that behave
predicatble and allows us to run some integration tests against this app
from a Rake task.

Note that the `rake integration_test` sends `kill -9` to the Sinatra
process after it's done. This is ungraceful and a bit nasty, but we don't
want zombie processes hanging around.

As part of this commit we've narrowed down the default test task to
*only* run the unit tests.

Also note that the "integration test" in this commit doesn't test any
part of our library - only that the test app is started correctly.